requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.5
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
random-user-agent==1.0.1
requests==2.32.2
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
zope.interface==6.4.post2
